# api/swagger.yaml
openapi: 3.0.0
info:
  title: Portfolio API
  description: Backend API for Kumar Nishant's Portfolio
  version: 1.1.0
servers:
  - url: http://localhost:5000/api
    description: Local Development Server
paths:
  /health:
    get:
      summary: Health Check
      description: Returns the status of the API server.
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /guestbook:
    get:
      summary: Get Guestbook Entries
      description: Retrieve guestbook entries. Email addresses and edit tokens are never exposed publicly.
      responses:
        '200':
          description: List of entries
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                    message:
                      type: string
                    avatar:
                      type: string
                      nullable: true
                    created_at:
                      type: string
        '429':
          description: Rate limit exceeded
    post:
      summary: Sign Guestbook
      description: Submit a new entry to the guestbook. Returns an edit token that is required to delete this entry later.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - message
              properties:
                name:
                  type: string
                  maxLength: 80
                message:
                  type: string
                  maxLength: 1000
                email:
                  type: string
                  nullable: true
                avatar:
                  type: string
                  nullable: true
      responses:
        '201':
          description: Entry created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  id:
                    type: integer
                  editToken:
                    type: string
        '400':
          description: Validation failed
        '429':
          description: Rate limit exceeded
  /guestbook/{id}:
    delete:
      summary: Delete Guestbook Entry
      description: |
        Deletes an entry. Authorization required via one of:
        - `Authorization: Bearer <editToken>` (the per-entry token returned when the entry was created), or
        - `x-admin-key: <ADMIN_API_KEY>` (server admin key)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Entry deleted
        '400':
          description: Invalid entry id
        '401':
          description: Missing authorization token
        '403':
          description: Token does not match this entry / admin key invalid
        '429':
          description: Rate limit exceeded
  /collaborate:
    post:
      summary: Submit Collaboration Request
      description: Sends a collaboration/inquiry email notification via Nodemailer.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
              properties:
                name:
                  type: string
                  maxLength: 80
                email:
                  type: string
                  format: email
                  maxLength: 120
                company:
                  type: string
                phone:
                  type: string
                projectType:
                  type: string
                budget:
                  type: string
                timeline:
                  type: string
                description:
                  type: string
                  maxLength: 2000
                requirements:
                  type: string
                  maxLength: 2000
      responses:
        '201':
          description: Collaboration request submitted
        '400':
          description: Validation failed
        '429':
          description: Rate limit exceeded